producer/requirements.txt
kafka-python==2.0.1
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
